pyproject.toml
[tool.poetry]
name = "lost-pets-api"
version = "0.1.0"
description = ""
authors = ["Wesley Mendes <wesley@wesleymendes.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.27"
psycopg2-binary = "^2.9.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.11b1"
flake8 = "^4.0.1"
ipdb = "^0.13.9"
pytest = "^6.2.5"
pytest-flake8 = "^1.0.7"
requests = "^2.26.0"
pytest-pythonpath = "^0.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
